requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
nltk==3.4
oauthlib==3.0.1
Pillow==5.4.1
PySocks==1.6.8
requests==2.21.0
requests-oauthlib==1.2.0
singledispatch==3.4.0.3
six==1.12.0
tweepy==3.7.0
urllib3==1.24.1
win-inet-pton==1.1.0
